+++ lib/net.dart
@@ -11,3 +11,5 @@ export 'src/net/req_email.dart';
 export 'src/net/req_ipv4.dart';
 export 'src/net/req_ipv6.dart';
 export 'src/net/req_mac_addr.dart';
+export 'src/net/req_url.dart';
+export 'src/net/url.dart';